chore: flatten protocols — delete abstraction layer (#322)
