--- silas/protocols/__init__.py
@@ -1,51 +1,0 @@
-from silas.protocols.approval import ApprovalManager, ApprovalVerifier, NonceStore
-from silas.protocols.audit import AuditLog
-from silas.protocols.channels import ChannelAdapterCore, RichCardChannel
-from silas.protocols.connections import ConnectionManager
-from silas.protocols.context import ContextManager
-from silas.protocols.execution import EphemeralExecutor, KeyManager, SandboxManager
-from silas.protocols.gates import GateCheckProvider, GateRunner
-from silas.protocols.goals import GoalManager
-from silas.protocols.memory import (
-    MemoryConsolidator,
-    MemoryPortability,
-    MemoryRetriever,
-    MemoryStore,
-)
-from silas.protocols.personality import PersonalityEngine, PersonaStore
-from silas.protocols.proactivity import AutonomyCalibrator, SuggestionEngine
-from silas.protocols.scheduler import TaskScheduler
-from silas.protocols.skills import SkillLoader, SkillResolver
-from silas.protocols.work import PlanParser, VerificationRunner, WorkItemExecutor, WorkItemStore
-
-__all__ = [
-    "ApprovalManager",
-    "ApprovalVerifier",
-    "AuditLog",
-    "AutonomyCalibrator",
-    "ChannelAdapterCore",
-    "ConnectionManager",
-    "ContextManager",
-    "EphemeralExecutor",
-    "GateCheckProvider",
-    "GateRunner",
-    "GoalManager",
-    "KeyManager",
-    "MemoryConsolidator",
-    "MemoryPortability",
-    "MemoryRetriever",
-    "MemoryStore",
-    "NonceStore",
-    "PersonaStore",
-    "PersonalityEngine",
-    "PlanParser",
-    "RichCardChannel",
-    "SandboxManager",
-    "SkillLoader",
-    "SkillResolver",
-    "SuggestionEngine",
-    "TaskScheduler",
-    "VerificationRunner",
-    "WorkItemExecutor",
-    "WorkItemStore",
-]
